requirements.txt
acme
josepy
OpenSSL
cryptography
azure-storage
